+++ components/tables/podcast/EpisodeRow.vue
@@ -15,7 +15,7 @@
 
       <p class="text-sm font-semibold">{{ title }}</p>
 
-      <p class="text-sm text-fg episode-subtitle mt-1.5 mb-0.5" v-html="subtitle" />
+      <p class="text-sm text-fg episode-subtitle mt-1.5 mb-0.5" v-html="$sanitizeHtml(subtitle)" />
 
       <p v-if="sortKey === 'audioFile.metadata.filename'" class="text-xs text-fg-muted truncate mt-2 mb-0.5">
         <span class="font-semibold">{{ $getString('LabelFilename') }}</span